Added pack classes with __init__.py witch contain CompEye and Mouse classes. Solved class divide problem.
